Rename user-config to userConfig
